Initial rewrite of the matrix driver
Split the singular driver in to a DMA-specific portion (which could be
refactored to support things like the ESP32-S2 and ESP32) and an RGB
matrix sync manager. There's still a lot of tests to write, and I expect
there to be bugs exposed by these tests, but we're getting there.